+++ include/powsybl/iidm/Generator.hpp
@@ -19,7 +19,6 @@ namespace powsybl {
 namespace iidm {
 
 class Terminal;
-class VariantManagerHolder;
 
 class Generator : public Injection, public ReactiveLimitsHolder {
 public:
