+++ include/powsybl/iidm/Load.hpp
@@ -17,8 +17,6 @@ namespace powsybl {
 
 namespace iidm {
 
-class VariantManagerHolder;
-
 class Load : public Injection {
 public:
     Load(VariantManagerHolder& network, const std::string& id, const std::string& name, bool fictitious, const LoadType& loadType,
