+++ src/iidm/ConnectedComponentsManager.cpp
@@ -10,8 +10,6 @@
 #include <powsybl/iidm/Bus.hpp>
 #include <powsybl/iidm/HvdcLine.hpp>
 #include <powsybl/iidm/Network.hpp>
-#include <powsybl/iidm/Terminal.hpp>
-#include <powsybl/iidm/VoltageLevel.hpp>
 #include <powsybl/stdcxx/make_unique.hpp>
 
 #include "ConnectedComponent.hpp"
