+++ src/iidm/extensions/LoadDetail.cpp
@@ -11,9 +11,7 @@
 
 #include <powsybl/PowsyblException.hpp>
 #include <powsybl/iidm/Load.hpp>
-#include <powsybl/iidm/Terminal.hpp>
 #include <powsybl/iidm/VariantManager.hpp>
-#include <powsybl/iidm/VoltageLevel.hpp>
 #include <powsybl/stdcxx/format.hpp>
 
 namespace powsybl {
